done all the changes, related to the cart and payment module

+++ server/v1/controller/cartController.js
@@ -1,13 +1,13 @@
 const cartService = require('../service/cartServices');
-
+const CONFIG = require('../../utils/appConfig')
 // Add item to cart
 exports.addItem = async (req, res) => {
     const { userId, productId, quantity } = req.body;
     try {
-        const results = await cartService.addItem(userId, productId, quantity);
-        res.status(201).json({ message: 'Item added to cart', results });
+        await cartService.addItem(userId, productId, quantity);
+        return res.success(CONFIG.SUCCESS_CODE, CONFIG.ITEM_ADDED_TO_CART)
     } catch (error) {
-        res.status(500).json({ error: error.message });
+        return res.reject(CONFIG.ERROR_CODE_INTERNAL_SERVER_ERROR, error.message)
     }
 };
 
@@ -17,11 +17,12 @@ exports.editItem = async (req, res) => {
     try {
         const results = await cartService.editItem(userId, productId, quantity);
         if (results.affectedRows === 0) {
-            return res.status(404).json({ message: 'Item not found in cart' });
+            // return res.status(404).json({ message: 'Item not found in cart' });
+            return res.success(CONFIG.SUCCESS_CODE, CONFIG.ITEM_NOT_FOUND_IN_CART)
         }
-        res.json({ message: 'Item quantity updated', results });
+        return res.success(CONFIG.SUCCESS_CODE, CONFIG.ITEM_QUANTITY_UPDATED)
     } catch (error) {
-        res.status(500).json({ error: error.message });
+        return res.reject(CONFIG.ERROR_CODE_INTERNAL_SERVER_ERROR, error.message);
     }
 };
 
@@ -29,12 +30,19 @@ exports.editItem = async (req, res) => {
 exports.deleteItem = async (req, res) => {
     const { userId, productId } = req.body;
     try {
-        const results = await cartService.deleteItem(userId, productId);
-        if (results.affectedRows === 0) {
-            return res.status(404).json({ message: 'Item not found in cart' });
+        const dataExists = await cartService.itemExist(userId, productId);
+        if (dataExists) {
+            const results = await cartService.deleteItem(userId, productId);
+            if (results) {
+                return res.success(CONFIG.SUCCESS_CODE, CONFIG.ITEM_REMOVED_FROM_CART)
+            } else {
+                return res.reject(CONFIG.ERROR_CODE_INTERNAL_SERVER_ERROR, CONFIG.INTERNAL_SERVER_ERROR)
+            }
+
+        } else {
+            return res.success(CONFIG.SUCCESS_CODE, CONFIG.ITEM_NOT_FOUND_IN_CART)
         }
-        res.json({ message: 'Item removed from cart', results });
     } catch (error) {
-        res.status(500).json({ error: error.message });
+        return res.reject(CONFIG.ERROR_CODE_INTERNAL_SERVER_ERROR, error.message);
     }
 };
